backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

python-dotenv==1.0.1
httpx==0.23.3   # ✅ compatible with supabase

python-multipart==0.0.9
supabase==1.0.3

# Document/file processing
python-docx==1.1.0
PyPDF2==3.0.1

aiofiles==23.2.1
sqlalchemy==2.0.30
transformer 
gdown